requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
mysqlclient
# psycopg2
passlib
python-jose[cryptography]